refactor: move migration_history logic from ghost sync to cutover
