+++ server/task_executor_schema_update_ghost_cutover.go
@@ -1,14 +1,22 @@
 package server
 
 import (
+	"bytes"
 	"context"
 	"encoding/json"
 	"fmt"
 	"os"
+	"strings"
 	"sync/atomic"
 	"time"
 
 	"github.com/bytebase/bytebase/api"
+	"github.com/bytebase/bytebase/common"
+	"github.com/bytebase/bytebase/common/log"
+	"github.com/bytebase/bytebase/plugin/db"
+	"github.com/bytebase/bytebase/plugin/db/util"
+	vcsPlugin "github.com/bytebase/bytebase/plugin/vcs"
+	"go.uber.org/zap"
 )
 
 // NewSchemaUpdateGhostCutoverTaskExecutor creates a schema update (gh-ost) cutover task executor.
@@ -19,6 +27,76 @@ func NewSchemaUpdateGhostCutoverTaskExecutor() TaskExecutor {
 // SchemaUpdateGhostCutoverTaskExecutor is the schema update (gh-ost) cutover task executor.
 type SchemaUpdateGhostCutoverTaskExecutor struct {
 	completed int32
+}
+
+func cutover(ctx context.Context, server *Server, task *api.Task, statement, schemaVersion string, vcsPushEvent *vcsPlugin.PushEvent, socketFilename, postponeFilename string) (terminated bool, result *api.TaskRunResultPayload, err error) {
+	statement = strings.TrimSpace(statement)
+
+	mi, err := preMigration(ctx, server, task, db.Migrate, statement, schemaVersion, vcsPushEvent)
+	if err != nil {
+		return true, nil, err
+	}
+	migrationID, schema, err := func() (migrationHistoryID int64, updatedSchema string, resErr error) {
+		driver, err := getAdminDatabaseDriver(ctx, task.Instance, task.Database.Name, "" /* pgInstanceDir */)
+		if err != nil {
+			return -1, "", err
+		}
+		defer driver.Close(ctx)
+		needsSetup, err := driver.NeedsSetupMigration(ctx)
+		if err != nil {
+			return -1, "", fmt.Errorf("failed to check migration setup for instance %q: %w", task.Instance.Name, err)
+		}
+		if needsSetup {
+			return -1, "", common.Errorf(common.MigrationSchemaMissing, fmt.Errorf("missing migration schema for instance %q", task.Instance.Name))
+		}
+
+		executor := driver.(util.MigrationExecutor)
+
+		var prevSchemaBuf bytes.Buffer
+		if _, err := driver.Dump(ctx, mi.Database, &prevSchemaBuf, true); err != nil {
+			return -1, "", err
+		}
+
+		insertedID, err := util.BeginMigration(ctx, executor, mi, prevSchemaBuf.String(), statement, db.BytebaseDatabase)
+		if err != nil {
+			if common.ErrorCode(err) == common.MigrationAlreadyApplied {
+				return insertedID, prevSchemaBuf.String(), nil
+			}
+			return -1, "", err
+		}
+		startedNs := time.Now().UnixNano()
+
+		defer func() {
+			if err := util.EndMigration(ctx, executor, startedNs, insertedID, updatedSchema, db.BytebaseDatabase, resErr == nil /*isDone*/); err != nil {
+				log.Error("failed to update migration history record",
+					zap.Error(err),
+					zap.Int64("migration_id", migrationHistoryID),
+				)
+			}
+		}()
+
+		if err := os.Remove(postponeFilename); err != nil {
+			return -1, "", fmt.Errorf("failed to remove postpone flag file, error: %w", err)
+		}
+
+		ticker := time.NewTicker(time.Second * 1)
+		defer ticker.Stop()
+
+		for range ticker.C {
+			if _, err := os.Stat(socketFilename); err != nil {
+				break
+			}
+		}
+
+		var afterSchemaBuf bytes.Buffer
+		if _, err := executor.Dump(ctx, mi.Database, &afterSchemaBuf, true /*schemaOnly*/); err != nil {
+			return -1, "", util.FormatError(err)
+		}
+
+		return insertedID, afterSchemaBuf.String(), nil
+	}()
+
+	return postMigration(ctx, server, task, vcsPushEvent, mi, migrationID, schema)
 }
 
 // RunOnce will run SchemaUpdateGhostCutover task once.
@@ -47,20 +125,7 @@ func (exec *SchemaUpdateGhostCutoverTaskExecutor) RunOnce(ctx context.Context, s
 	socketFilename := getSocketFilename(syncTaskID, task.Database.ID, task.Database.Name, tableName)
 	postponeFilename := getPostponeFlagFilename(syncTaskID, task.Database.ID, task.Database.Name, tableName)
 
-	if err := os.Remove(postponeFilename); err != nil {
-		return true, nil, fmt.Errorf("failed to remove postpone flag file, error: %w", err)
-	}
-
-	ticker := time.NewTicker(time.Second * 1)
-	defer ticker.Stop()
-
-	for range ticker.C {
-		if _, err := os.Stat(socketFilename); err != nil {
-			break
-		}
-	}
-
-	return true, &api.TaskRunResultPayload{Detail: "cutover done"}, nil
+	return cutover(ctx, server, task, payload.Statement, payload.SchemaVersion, payload.VCSPushEvent, socketFilename, postponeFilename)
 }
 
 // IsCompleted tells the scheduler if the task execution has completed
